# api-contract/src/main/resources/static/brightlife_api_contract.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Bright-life Enhancement API
  description:
    "API documentation for Bright-life Enhancement project"
  license:
    name: "fronteersdev Inc"
    url: "https://fronteersdev.com/license"
servers:
  - url: http://localhost:7081
    description: Local test server
  - url: https://bright-life-enhancement.fronteers-dev.com/
    description: Dev test server
security:
  - BearerAuth: [ ]  # ⬅️ Apply JWT security globally to all endpoints
components:
  securitySchemes:
    BearerAuth:
      type: http
      description: This API uses JWT for authentication
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Passport:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
          format: int32
        file:
          $ref: '#/components/schemas/FileUpload'
    PassportResponse:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
          format: int32
        passportUrl:
          type: string
...
    FileUpload:
      type: string
      format: binary
...
  responses:
    BadRequest:
      #      type: object
      description: |
        Bad Request
        The request body or parameters does not conform to the schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        Unauthorized.

        - Indicates the request does not contain a valid token to authenticate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Forbidden.
            - Indicates the authenticated entity does not have the privilege to perform
              the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnProcessableEntity:
      description: |
        Unprocessable entity.
        - The request body may be well-formed, but either request payload or parameters failed semantic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: |
        Internal Server Error.
            - The server encountered an unexpected error while processing the request..
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /api/v1/patients/passports/upload:
    post:
      summary:
        Uploads patient passport to s3bucket
      tags:
        - Patients
      operationId: uploadPatientPassport
      requestBody:
        description: Uploads Passport
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Passport"
      responses:
        "200":
          description: Passport Uploaded Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/search:
    post:
      summary:
        Searches and returns all Patients matching all of the search criteria fields passed.
        The number search results shall be constrained by the limit params in Query.
        Search criteria, applies an **AND** based search on all fields passed in the request.
        If the search criteria is empty/null, it returns list of patient
      tags:
        - Patients
      operationId: listPatients
      parameters:
        - description: The page number (starting from 1)
          $ref: "#/components/parameters/PageParam"
        - description: How many items to return at one time (max 100)
          $ref: "#/components/parameters/limitParam"
      requestBody:
        description:
          "Search criteria, applies an **AND** based search on all fields\
                \ passed in the request."
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientSearch"
      responses:
        "200":
          description: Patients Records Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPatients"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/{patientId}:
    get:
      summary: Fetches A Patient's Details
      tags:
        - Patients
      operationId: getPatient
      parameters:
        - $ref: "#/components/parameters/patientId"
      responses:
        "200":
          description: Patient Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/patients/forms/register/{patientId}:
    get:
      summary: Fetches A Patients Registration Details
      tags:
        - Patients
      operationId: getRegistration
      parameters:
        - $ref: "#/components/parameters/patientId"
      responses:
        "200":
          description: Patient Registration Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientRegistrationForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/adhd:
    post:
      summary: Submits A New ADHD Form For A Patient
      tags:
        - Patients
      operationId: submitAdhd
      requestBody:
        description: This is used to submit a new ADHD form an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ADHDForm"
      responses:
        "201":
          description: Patient ADHD Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/adhd/{idParam}:
    get:
      summary: Fetches A Patients ADHD Form Details
      tags:
        - Patients
      operationId: getAdhd
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient ADHD Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ADHDForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/anxiety-disorder:
    post:
      summary: Submits A New Anxiety Disorder Form For A Patient
      tags:
        - Patients
      operationId: submitAnxietyDisorder
      requestBody:
        description: This is used to submit a new Anxiety Disorder form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnxietyDisorderForm"
      responses:
        "201":
          description: Patient Anxiety Disorder Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/anxiety-disorder/{idParam}:
    get:
      summary: Fetches A Patients Anxiety Disorder Form Details
      tags:
        - Patients
      operationId: getAnxietyDisorder
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Anxiety Disorder Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnxietyDisorderForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/controlled-substance:
    post:
      summary: Submits A New Controlled Substance Form For A Patient
      tags:
        - Patients
      operationId: submitControlledSubstance
      requestBody:
        description: This is used to submit a new Controlled Substance form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlledSubstanceForm"
      responses:
        "201":
          description: Patient Controlled Substance Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/controlled-substance/{idParam}:
    get:
      summary: Fetches A Patients Controlled Substance Form Details
      tags:
        - Patients
      operationId: getControlledSubstance
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Controlled Substance Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlledSubstanceForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/depression-assessment:
    post:
      summary: Submits A New DepressionAssessment Form For A Patient
      tags:
        - Patients
      operationId: submitDepressionAssessment
      requestBody:
        description: This is used to submit a new Depression Assessment Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepressionAssessmentForm"
      responses:
        "201":
          description: Patient Depression Assessment Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/depression-assessment/{idParam}:
    get:
      summary: Fetches A Patient's Depression Assessment Form Details
      tags:
        - Patients
      operationId: getDepressionAssessment
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Controlled Substance Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepressionAssessmentForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/initial-evaluation:
    post:
      summary: Submits A New Initial Evaluation Form For A Patient
      tags:
        - Patients
      operationId: submitInitialEvaluation
      requestBody:
        description: This is used to submit a new Initial Evaluation Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitialEvaluationForm"
      responses:
        "201":
          description: Patient Initial Evaluation Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/initial-evaluation/{idParam}:
    get:
      summary: Fetches A Patient's Initial Evaluation Form Details
      tags:
        - Patients
      operationId: getInitialEvaluation
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Initial Evaluation Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitialEvaluationForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/intake:
    post:
      summary: Submits A New Intake Form For A Patient
      tags:
        - Patients
      operationId: submitIntake
      requestBody:
        description: This is used to submit a new Intake Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeForm"
      responses:
        "201":
          description: Patient Intake Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/intake/{idParam}:
    get:
      summary: Fetches A Patient's Intake Form Details
      tags:
        - Patients
      operationId: getIntake
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Intake Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/medication-consent:
    post:
      summary: Submits A New Medication Consent Form For A Patient
      tags:
        - Patients
      operationId: submitMedicationConsent
      requestBody:
        description: This is used to submit a new Medication Consent Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationConsentForm"
      responses:
        "201":
          description: Patient Medication Consent Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/medication-consent/{idParam}:
    get:
      summary: Fetches A Patient's Medication Consent Form Details
      tags:
        - Patients
      operationId: getMedicationConsent
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Medication Consent Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationConsentForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/mood-disorder-assessment:
    post:
      summary: Submits A New Mood Disorder Assessment Form For A Patient
      tags:
        - Patients
      operationId: submitMoodDisorderAssessment
      requestBody:
        description: This is used to submit a new MoodDisorderAssessment Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodDisorderAssessmentForm"
      responses:
        "201":
          description: Patient Mood Disorder Assessment Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/mood-disorder-assessment/{idParam}:
    get:
      summary: Fetches A Patient's Mood Disorder Assessment Form Details
      tags:
        - Patients
      operationId: getMoodDisorderAssessment
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient Mood Disorder Assessment Form Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodDisorderAssessmentForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/privacy-practices:
    post:
      summary: Submits A New Notice Of Privacy Practices Form For A Patient
      tags:
        - Patients
      operationId: submitNoticeOfPrivacyPractices
      requestBody:
        description: This is used to submit a new NoticeOfPrivacyPractices Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeOfPrivacyPracticesForm"
      responses:
        "201":
          description: Patient NoticeOfPrivacyPractices Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/privacy-practices/{idParam}:
    get:
      summary: Fetches A Patient's Notice Of Privacy Practices Assessment Form Details
      tags:
        - Patients
      operationId: getNoticeOfPrivacyPractices
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient NoticeOfPrivacyPractices Form Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeOfPrivacyPracticesForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/info-and-fin-policy:
    post:
      summary: Submits A Patient Information Consent And Financial Policy Form For A Patient
      tags:
        - Patients
      operationId: submitPatientInformationConsentAndFinancialPolicy
      requestBody:
        description: This is used to submit a new PatientInformationConsentAndFinancialPolicy Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInformationConsentAndFinancialPolicyForm"
      responses:
        "201":
          description: Patient PatientInformationConsentAndFinancialPolicy Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/info-and-fin-policy/{idParam}:
    get:
      summary: Fetches A Patient's Information Consent And Financial Policy Form Details
      tags:
        - Patients
      operationId: getPatientInformationConsentAndFinancialPolicy
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient PatientInformationConsentAndFinancialPolicy Form Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientInformationConsentAndFinancialPolicyForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/release-receive:
    post:
      summary: Submits A Release Receive Form For A Patient
      tags:
        - Patients
      operationId: submitReleaseReceive
      requestBody:
        description: This is used to submit a new ReleaseReceive Form for an already existing patient
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseReceiveForm"
      responses:
        "201":
          description: Patient ReleaseReceive Form Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/patients/forms/release-receive/{idParam}:
    get:
      summary: Fetches A Patient's ReleaseReceive Form Details
      tags:
        - Patients
      operationId: getReleaseReceive
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Patient ReleaseReceive Form Details Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseReceiveForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...